Replace open quote with apostrophe
We aren't opening a quotation here, we just need an apostrophe.  Replace
it with "'", the straight apostrophe that also doubles as a straight
quote; wile we could use the proper curly quote for it, "’", none of the
rest of the apostrophes use that for, so for consistency just use the
straight apostrophe from the ASCII range.