Replace parentheses with code tags
In one example, it appears that the original paper uses parentheses to
make it explicit where a code sample in running text starts and ends.
Since we now have the ability to use the code tag for the same purpose,
remove the redundant parentheses (and properly tag one of the examples).