Add vertical bars to "length x" alternative
Earlier in the text, Pratt referrs to "|x|" as an alternative to "length
x", as "length x" wouldn't work properly in the type hierarchy that is
used to determine associativity.  Later on, he mentions that the
programmer 'must know that one writes " x ", not "length x"', which
seems to be a clear typo in which the vertical bars were not added
in (although the space for them had been left).